augmented the AjaxException class to accept a default URL query part (reduces code duplication in a lot of places) + augmented various backend sources to better use the exception class (and its 'automated' redirection-on-error behaviour): this way the admin pages behave more appropriately under (severe) error conditions.

+++ lib/modules/news/news.Process.php
@@ -80,54 +80,62 @@ $do_action 	= getGETparam4IdOrNumber('action');
  */
 if($_SERVER['REQUEST_METHOD'] == 'POST' && $do_action == 'add-edit-news' && checkAuth()) 
 {
-	FbX::SetFeedbackLocation($cfg['rootdir'] . 'lib/modules/news/news.Manage.php');
-	
+	FbX::SetFeedbackLocation('news.Manage.php');
 	try
 	{
-		// Only if current user has the rights
-		if($perm['manageModNews']>0 && $_SESSION['ccms_userLevel']>=$perm['manageModNews']) 
+		if (!empty($pageID))
 		{
-			// Published
-			$newsAuthor = getPOSTparam4Number('newsAuthor');
-			//$pageID = getPOSTparam4IdOrNumber('pageID');
-			$newsTitle = getPOSTparam4DisplayHTML('newsTitle');
-			$newsTeaser = getPOSTparam4DisplayHTML('newsTeaser');
-			$newsContent = getPOSTparam4DisplayHTML('newsContent');
-			$newsModified = getPOSTparam4DateTime('newsModified', time());
-			$newsPublished = getPOSTparam4boolean('newsPublished');
+			FbX::SetFeedbackLocation('news.Manage.php', 'file=' . $pageID);
+		
+			// Only if current user has the rights
+			if($perm['manageModNews']>0 && $_SESSION['ccms_userLevel']>=$perm['manageModNews']) 
+			{
+				// Published
+				$newsAuthor = getPOSTparam4Number('newsAuthor');
+				//$pageID = getPOSTparam4IdOrNumber('pageID');
+				$newsTitle = getPOSTparam4DisplayHTML('newsTitle');
+				$newsTeaser = getPOSTparam4DisplayHTML('newsTeaser');
+				$newsContent = getPOSTparam4DisplayHTML('newsContent');
+				$newsModified = getPOSTparam4DateTime('newsModified', time());
+				$newsPublished = getPOSTparam4boolean('newsPublished');
 
-			/* make sure empty news posts don't get through! front-end checking alone is NOT enough! */
-			if (!empty($pageID) && !empty($newsAuthor) && !empty($newsAuthor) && !empty($newsTitle) && strlen($newsTitle) >= 3 && !empty($newsTeaser) && !empty($newsContent))
-			{
-				// Set all the submitted variables
-				$values = array(); // [i_a] make sure $values is an empty array to start with here
-				$values["userID"] = MySQL::SQLValue($newsAuthor, MySQL::SQLVALUE_NUMBER);
-				$values["pageID"] = MySQL::SQLValue($pageID, MySQL::SQLVALUE_TEXT);
-				$values["newsTitle"] = MySQL::SQLValue($newsTitle, MySQL::SQLVALUE_TEXT);
-				$values["newsTeaser"] = MySQL::SQLValue($newsTeaser, MySQL::SQLVALUE_TEXT);
-				$values["newsContent"] = MySQL::SQLValue($newsContent, MySQL::SQLVALUE_TEXT);
-				$values["newsModified"] = MySQL::SQLValue($newsModified, MySQL::SQLVALUE_DATETIME);
-				$values["newsPublished"] = MySQL::SQLValue($newsPublished, MySQL::SQLVALUE_BOOLEAN);
-			
-				// Execute either INSERT or UPDATE based on $newsID
-				if($db->AutoInsertUpdate($cfg['db_prefix'].'modnews', $values, array('newsID' => MySQL::BuildSQLValue($newsID)))) 
+				/* make sure empty news posts don't get through! front-end checking alone is NOT enough! */
+				if (!empty($pageID) && !empty($newsAuthor) && !empty($newsAuthor) && !empty($newsTitle) && strlen($newsTitle) >= 3 && !empty($newsTeaser) && !empty($newsContent))
 				{
-					header('Location: ' . makeAbsoluteURI('news.Manage.php?file='.$pageID.'&status=notice&msg='.rawurlencode($ccms['lang']['backend']['itemcreated'])));
-					exit();
-				} 
-				else 
-				{
-					throw new FbX($db->MyDyingMessage(), 'file=' . $pageID);
+					// Set all the submitted variables
+					$values = array(); // [i_a] make sure $values is an empty array to start with here
+					$values["userID"] = MySQL::SQLValue($newsAuthor, MySQL::SQLVALUE_NUMBER);
+					$values["pageID"] = MySQL::SQLValue($pageID, MySQL::SQLVALUE_TEXT);
+					$values["newsTitle"] = MySQL::SQLValue($newsTitle, MySQL::SQLVALUE_TEXT);
+					$values["newsTeaser"] = MySQL::SQLValue($newsTeaser, MySQL::SQLVALUE_TEXT);
+					$values["newsContent"] = MySQL::SQLValue($newsContent, MySQL::SQLVALUE_TEXT);
+					$values["newsModified"] = MySQL::SQLValue($newsModified, MySQL::SQLVALUE_DATETIME);
+					$values["newsPublished"] = MySQL::SQLValue($newsPublished, MySQL::SQLVALUE_BOOLEAN);
+				
+					// Execute either INSERT or UPDATE based on $newsID
+					if($db->AutoInsertUpdate($cfg['db_prefix'].'modnews', $values, array('newsID' => MySQL::BuildSQLValue($newsID)))) 
+					{
+						header('Location: ' . makeAbsoluteURI('news.Manage.php?file='.$pageID.'&status=notice&msg='.rawurlencode($ccms['lang']['backend']['itemcreated'])));
+						exit();
+					} 
+					else 
+					{
+						throw new FbX($db->MyDyingMessage());
+					}
 				}
-			}
-			else
+				else
+				{
+					throw new FbX($ccms['lang']['system']['error_pagetitle']);
+				}
+			} 
+			else 
 			{
-				throw new FbX($ccms['lang']['system']['error_pagetitle'], 'file=' . $pageID);
+				throw new FbX($ccms['lang']['auth']['featnotallowed']);
 			}
 		} 
 		else 
 		{
-			throw new FbX($ccms['lang']['auth']['featnotallowed'], 'file=' . $pageID);
+			throw new FbX($ccms['lang']['auth']['error_forged']);
 		}
 	}
 	catch (CcmsAjaxFbException $e)
@@ -147,45 +155,54 @@ if($_SERVER['REQUEST_METHOD'] == 'POST' && $do_action == 'del-news' && checkAuth
 	
 	try
 	{
-		// Only if current user has the rights
-		if($perm['manageModNews']>0 && $_SESSION['ccms_userLevel']>=$perm['manageModNews']) 
+		if (!empty($pageID))
 		{
-			// Number of selected items
-			$total = (!empty($_POST['newsID']) && is_array($_POST['newsID']) ? count($_POST['newsID']) : 0);
-			
-			// If nothing selected, throw error
-			if($total==0) 
-			{
-				throw new FbX($ccms['lang']['system']['error_selection'], 'file=' . $pageID);
-			}
-			
-			// Delete details from the database
-			$i=0;
-			foreach ($_POST['newsID'] as $idnum) 
-			{
-				$idnum = filterParam4Number($idnum);
-				
-				$values = array(); // [i_a] make sure $values is an empty array to start with here
-				$values['newsID'] = MySQL::SQLValue($idnum, MySQL::SQLVALUE_NUMBER);
-				$result = $db->DeleteRows($cfg['db_prefix'].'modnews', $values);
-				if (!$result) break;
-				$i++;
-			}
+			FbX::SetFeedbackLocation('news.Manage.php', 'file=' . $pageID);
 		
-			// Check for errors
-			if($result && $i==$total) 
+			// Only if current user has the rights
+			if($perm['manageModNews']>0 && $_SESSION['ccms_userLevel']>=$perm['manageModNews']) 
 			{
-				header('Location: ' . makeAbsoluteURI('news.Manage.php?file='.$pageID.'&status=notice&msg='.rawurlencode($ccms['lang']['backend']['fullremoved'])));
-				exit();
+				// Number of selected items
+				$total = (!empty($_POST['newsID']) && is_array($_POST['newsID']) ? count($_POST['newsID']) : 0);
+				
+				// If nothing selected, throw error
+				if($total==0) 
+				{
+					throw new FbX($ccms['lang']['system']['error_selection']);
+				}
+				
+				// Delete details from the database
+				$i=0;
+				foreach ($_POST['newsID'] as $idnum) 
+				{
+					$idnum = filterParam4Number($idnum);
+					
+					$values = array(); // [i_a] make sure $values is an empty array to start with here
+					$values['newsID'] = MySQL::SQLValue($idnum, MySQL::SQLVALUE_NUMBER);
+					$result = $db->DeleteRows($cfg['db_prefix'].'modnews', $values);
+					if (!$result) break;
+					$i++;
+				}
+			
+				// Check for errors
+				if($result && $i==$total) 
+				{
+					header('Location: ' . makeAbsoluteURI('news.Manage.php?file='.$pageID.'&status=notice&msg='.rawurlencode($ccms['lang']['backend']['fullremoved'])));
+					exit();
+				} 
+				else 
+				{
+					throw new FbX($db->MyDyingMessage());
+				}
 			} 
 			else 
 			{
-				throw new FbX($db->MyDyingMessage(), 'file=' . $pageID);
+				throw new FbX($ccms['lang']['auth']['featnotallowed']);
 			}
 		} 
 		else 
 		{
-			throw new FbX($ccms['lang']['auth']['featnotallowed'], 'file=' . $pageID);
+			throw new FbX($ccms['lang']['auth']['error_forged']);
 		}
 	}
 	catch (CcmsAjaxFbException $e)
@@ -201,21 +218,22 @@ if($_SERVER['REQUEST_METHOD'] == 'POST' && $do_action == 'del-news' && checkAuth
  */
 if($_SERVER['REQUEST_METHOD'] == 'POST' && $do_action == 'cfg-news' && checkAuth()) 
 {
-	FbX::SetFeedbackLocation($cfg['rootdir'] . 'lib/modules/news/news.Manage.php');
-	
+	FbX::SetFeedbackLocation('news.Manage.php');
 	try
 	{
-		// Only if current user has the rights
-		if($perm['manageModNews']>0 && $_SESSION['ccms_userLevel']>=$perm['manageModNews']) 
+		if (!empty($pageID))
 		{
-			$showLocale = getPOSTparam4IdOrNumber('locale');
-			$showMessage = getPOSTparam4Number('messages');
-			$showAuthor = getPOSTparam4boolean('author');
-			$showDate = getPOSTparam4boolean('show_modified');
-			$showTeaser = getPOSTparam4boolean('show_teaser');
-
-			if (!empty($pageID))
+			FbX::SetFeedbackLocation('news.Manage.php', 'file=' . $pageID);
+		
+			// Only if current user has the rights
+			if($perm['manageModNews']>0 && $_SESSION['ccms_userLevel']>=$perm['manageModNews']) 
 			{
+				$showLocale = getPOSTparam4IdOrNumber('locale');
+				$showMessage = getPOSTparam4Number('messages');
+				$showAuthor = getPOSTparam4boolean('author');
+				$showDate = getPOSTparam4boolean('show_modified');
+				$showTeaser = getPOSTparam4boolean('show_teaser');
+
 				$values = array(); // [i_a] make sure $values is an empty array to start with here
 				$values["pageID"]		= MySQL::SQLValue($pageID, MySQL::SQLVALUE_TEXT);
 				$values["showLocale"]	= MySQL::SQLValue($showLocale, MySQL::SQLVALUE_TEXT);
@@ -232,17 +250,17 @@ if($_SERVER['REQUEST_METHOD'] == 'POST' && $do_action == 'cfg-news' && checkAuth
 				} 
 				else 
 				{
-					throw new FbX($db->MyDyingMessage(), 'file=' . $pageID);
+					throw new FbX($db->MyDyingMessage());
 				}
-			}
+			} 
 			else 
 			{
-				throw new FbX($ccms['lang']['system']['error_forged'], 'file=' . $pageID);
+				throw new FbX($ccms['lang']['auth']['featnotallowed']);
 			}
-		} 
+		}
 		else 
 		{
-			throw new FbX($ccms['lang']['auth']['featnotallowed'], 'file=' . $pageID);
+			throw new FbX($ccms['lang']['system']['error_forged']);
 		}
 	}
 	catch (CcmsAjaxFbException $e)
